1825 - Moved utils
