+++ projects/_01_ingest/capacity_tracker/tests/test_ingest_capacity_tracker_data.py
@@ -43,8 +43,8 @@ class IngestCapacityTrackerDataTests(SparkBaseTest):
             NonResData.expected_rows, NonResSchema.sample_schema
         )
 
-    @patch(f"{PATCH_PATH}.read_partial_csv_content")
-    @patch(f"{PATCH_PATH}.get_s3_objects_list")
+    @patch(f"{PATCH_PATH}.iUtils.read_partial_csv_content")
+    @patch(f"{PATCH_PATH}.iUtils.get_s3_objects_list")
     @patch(f"{PATCH_PATH}.utils.write_to_parquet")
     @patch(f"{PATCH_PATH}.utils.read_csv")
     def test_main_runs_when_source_is_csv(
@@ -72,8 +72,8 @@ class IngestCapacityTrackerDataTests(SparkBaseTest):
             write_to_parquet_patch.call_args[0][1], self.TEST_NEW_DESTINATION
         )
 
-    @patch(f"{PATCH_PATH}.read_partial_csv_content")
-    @patch(f"{PATCH_PATH}.get_s3_objects_list")
+    @patch(f"{PATCH_PATH}.iUtils.read_partial_csv_content")
+    @patch(f"{PATCH_PATH}.iUtils.get_s3_objects_list")
     @patch(f"{PATCH_PATH}.utils.write_to_parquet")
     @patch(f"{PATCH_PATH}.utils.read_csv")
     def test_main_runs_when_source_is_directory(
