+++ projects/_01_ingest/cqc_pir/jobs/ingest_cqc_pir_data.py
@@ -4,35 +4,29 @@ import sys
 
 os.environ["SPARK_VERSION"] = "3.5"
 
+from projects._01_ingest.utils import utils as iUtils
 from schemas.cqc_pir_schema import PIR_SCHEMA
-from utils import utils
-from utils.s3_file_utils import (
-    construct_destination_path,
-    construct_s3_uri,
-    get_s3_objects_list,
-    is_csv,
-    split_s3_uri,
-)
+from utils import s3_file_utils, utils
 
 
 def main(source, destination):
-    if is_csv(source):
+    if iUtils.is_csv(source):
         print("Single file provided to job. Handling single file.")
-        bucket, key = split_s3_uri(source)
-        new_destination = construct_destination_path(destination, key)
+        bucket, key = s3_file_utils.split_s3_uri(source)
+        new_destination = iUtils.construct_destination_path(destination, key)
         ingest_pir_dataset(source, new_destination, PIR_SCHEMA)
         return
 
     print("Multiple files provided to job. Handling each file...")
-    bucket, prefix = split_s3_uri(source)
-    objects_list = get_s3_objects_list(bucket, prefix)
+    bucket, prefix = s3_file_utils.split_s3_uri(source)
+    objects_list = iUtils.get_s3_objects_list(bucket, prefix)
 
     print("Objects list:")
     print(objects_list)
 
     for key in objects_list:
-        new_source = construct_s3_uri(bucket, key)
-        new_destination = construct_destination_path(destination, key)
+        new_source = iUtils.construct_s3_uri(bucket, key)
+        new_destination = iUtils.construct_destination_path(destination, key)
         ingest_pir_dataset(new_source, new_destination, PIR_SCHEMA)
